[FIX] Treat attachments with content-type application/pgp-encrypted as encrypted attachments [#934 state:fixed]
In PGP/MIME, application/pgp-encrypted is used as the content type for the PGP/MIME version marker.
Some plugins or MUAs out there, seem to deem it appropriate however, to use the application/pgp-encrypted content-type for any encrypted attachment.

So in order to treat those attachments properly, GPGMail checks for the Version: 1 marker.