[FIX] Properly display pseudo-PGP/MIME messages from iPGMail [#938 state:fixed]
Due to restrictions of iOS, iPGMail is not able to create a proper PGP/MIME message structure.
The structure it currently creates looks like the following:

        multipart/mixed
            text/plain
            application/pgp-encrypted
            text/plain

In addition, if other encrypted files are attached as well the structure above is
extended with the following parts:

            application/octet-stream
            text/plain
            application/octet-stream
            text/plain
            …

As a result, GPGMail believes that this is a PGP/MIME structure, based on the
application/pgp-encrypted part, which contains the version in proper
PGP/MIME messages, and fails to handle the message.

To properly interpret iPGMail messages, GPGMail now handles the application/pgp-encrypted
part just like any other text/plain part containing encrypted data, if the BEGIN PGP MESSAGE
marker contains a Version: iPGMail header.