[FIX] Recognize signed attachments without detached-signature [#936 state:fixed]
In the past, GPGMail could only recognize signed attachments with a detached-signature, but displayed signed attachments with no detached-signature as simple attachments.

To fix that, GPGMail now checks if a signature is found on the attachment, and if no corresponding signed part is found, it will verify the attachment itself.