Fix regression introduced in fix for bug #903 [#903]
In order to support the gnupg.conf group option, GPGMail stored a list of GPGKey instances for each recipient.
The method checking if PGP should be used for encryption didn’t know about the list, since it only checked for GPGKey instances and not NSArray with GPGKey instances and thus assumed that S/MIME should be used for encryption.