[FIX] Only flag attachments as signed if they are in fact signed [#942 state:fixed]
Attachments that are either signed nor encrypted are falsely detected as signed, since GPGPacket returns a signature packet in some cases (e.g. some PNG files appear to have binary data which is interpreted as a pgp signature packet with an invalid version 71).
In order to avoid that, an attachment is only checked for signature packets, if the extension of the filename matches either a known encrypted or signed extension.