[FIX] Make sure GPGMail preference panel is only added once - second attempt [#943 state:fixed]
Due to an implementation error in the timing when the preference panel of GPGMail was registered with Mail.app it was possible for the preference panel to be added twice.
This bug was triggered if a second plugin was loaded prior to GPGMail.
The preference panel is now registered in +[MailPreferences makePreferenceTabViewItems] only if it was not previously registered.