[FIX] Properly recognize and discard OpenPGP keys during S/MIME certificate check [#962]
In order to support the gnupg group option, OpenPGP keys mapped to a recipient are no longer single GPGKey instances but rather always an NSArray of GPGKey instances. In order for the OpenPGP keys to be properly discarded, when checking for matching S/MIME certficates, the check has to take the fact into account, that it’s no longer enough to check for GPGKey instances, but rather for NSArray instances as well.